requirements/local.txt
-r ./base.txt
django-colorful
django-colorfield
django_localflavor
Werkzeug==0.15.2  # https://github.com/pallets/werkzeug
ipdb==0.12  # https://github.com/gotcha/ipdb
Sphinx==2.0.0  # https://github.com/sphinx-doc/sphinx
psycopg2-binary==2.7.7  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.670  # https://github.com/python/mypy
pytest==4.4.0  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.5  # https://github.com/PyCQA/flake8
coverage==4.5.3  # https://github.com/nedbat/coveragepy
black==19.3b0  # https://github.com/ambv/black
pylint-django==2.0.6  # https://github.com/PyCQA/pylint-django

# Django
# ------------------------------------------------------------------------------
factory-boy==2.11.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==1.11  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.1.6  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.4.8  # https://github.com/pytest-dev/pytest-django
